UefiCpuPkg/CpuPageTableLib: Add a new API to create new writable pagetable
Add a new API into CpuPageTableLib to create a new pagetable which
reuses parts of input origin page table and remaps a target range.
All present ranges in target range are marked as RW in this new page
table. The new pagetable is 1:1 mapping and this API assumes input
pagetable is 1:1 mapping.

Signed-off-by: Dun Tan <dun.tan@intel.com>
Cc: Eric Dong <eric.dong@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Rahul Kumar <rahul1.kumar@intel.com>