UefiCpuPkg: Avoid align 4096 in ExceptionHandler code
FRED spec need the FRED handler entrypoint for ring3 is 4K align, and current implementation use
"ALIGN   4096" to make sure this.
However, if the code section in this library is 4K align, it need the whole module which link
this library should be 4K align. This brings some issues:
1.Need add linker parameter to assign the alignemt is 4K, which needs to be done in platform dsc files
2.Need to make module location in Flash 4K alignment if the module links this library and run XIP.
3.All sections follow the same 4k alignment, and the module size will be larger

To fix this issue, add several 32-byte code block, each of them may becomes the FRED handler entry, and they
can jump to the next real FRED handler. This way, only 32 bytes alignment is needed

Cc: Eric Dong <eric.dong@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Rahul Kumar <rahul1.kumar@intel.com>
Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>