UefiCpuPkg: Support both FRED and IDT in CpuStackGuard test.
In CpuStackGuard test of Pei/DxeCpuExceptionHandlerLib, by
default, seperate stack is prepared only for DB in FRED, but
both PF and DB in Idt. To support both FRED and Idt, register
CpuStackGuardExceptionHandler for both PF and DF. Besides,
it is possible that both PF and DF are triggered in this test
case. Only first exception happens in CpuStackGuard test case
should write to mRspAddress to record the Rsp where exception
happens.

Signed-off-by: Dun Tan <dun.tan@intel.com>