UefiCpuPkg/PiSmmCpuDxeSmm: Split SMRAM Paging to 4KB
This patch is to split SMRAM Paging to 4KB during
SMM
initialization (SmmInitPageTable) so as to avoid
the SMRAM
paging-structure change when SMI happen
(PerformRemainingTasks).
The reason is to avoid the
Paging-Structure change impact
to the multiple
Processors. Refer SDM section "4.10.4" & "4.10.5".

Currently, SMM BSP needs to update the SMRAM
Paging-Structure
according MemoryAttributesTable
when SMM ready to lock happen.
Meanwhile, all APs
are still running in SMI, which might access
the
affected linear-address range between the time
of modification
and the time of invalidation access.
That will be a potential
problem leading the exception
happen.

This patch is to avoid the SMRAM Paging-Structure
change in later
SMI phase.

Signed-off-by: Dun Tan <dun.tan@intel.com>